pyproject.toml
[tool.poetry]
name = "craftconnect"
version = "0.1.0"
description = "A platform that connects artisans with customers, providing a marketplace for artisanal goods and services."
authors = ["Cozyamy <cozyamy@gmail.com>", "miraclemenikelechi",]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
pydantic = {extras = ["email"], version = "^2.7.0"}
phonenumbers = "^8.13.34"
pydantic-extra-types = "^2.6.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
firebase-admin = "^6.5.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
